feat: update via cli to start the verifier and the coordinator
