+++ infrastructure/via/src/token.ts
@@ -1,12 +1,37 @@
-import * as fs from 'fs';
 import { Command } from 'commander';
 import { Wallet, Provider, Contract } from 'zksync-ethers';
 import { ethers } from 'ethers';
+import * as utils from 'utils';
+
+const DEFAULT_DEPOSITOR_PRIVATE_KEY = 'cVZduZu265sWeAqFYygoDEE1FZ7wV9rpW5qdqjRkUehjaUMWLT1R';
+const DEFAULT_NETWORK = 'regtest';
+const DEFAULT_RPC_URL = 'http://0.0.0.0:18443';
+const DEFAULT_RPC_USERNAME = 'rpcuser';
+const DEFAULT_RPC_PASSWORD = 'rpcpassword';
 
 // 0x36615Cf349d7F6344891B1e7CA7C72883F5dc049
 const DEFAULT_L2_PRIVATE_KEY = '0x7726827caac94a7f9e1b160f7ea819f172f7b6f9d2a97f992c38edeab82d4110';
 const DEFAULT_L2_RPC_URL = 'http://0.0.0.0:3050';
 const L2_BASE_TOKEN = '0x000000000000000000000000000000000000800a';
+
+async function deposit(
+    amount: number,
+    receiverL2Address: string,
+    senderPrivateKey: string,
+    network: String,
+    rcpUrl: string,
+    rpcUsername: string,
+    rpcPassword: string
+) {
+    if (isNaN(amount)) {
+        console.error('Error: Invalid deposit amount. Please provide a valid number.');
+        return;
+    }
+    process.chdir(`${process.env.VIA_HOME}`);
+    await utils.spawn(
+        `cargo run --example deposit -- ${amount} ${receiverL2Address} ${senderPrivateKey} ${network} ${rcpUrl} ${rpcUsername} ${rpcPassword}`
+    );
+}
 
 async function withdraw(amount: number, receiverL1Address: string, userL2PrivateKey: string, rpcUrl: string) {
     if (isNaN(amount)) {
@@ -62,7 +87,28 @@ async function withdraw(amount: number, receiverL1Address: string, userL2Private
     console.log('Balance after withdraw', ethers.formatUnits(balance, 8));
 }
 
-export const command = new Command('token').description('Bridge BTC L2>L1');
+export const command = new Command('token').description('Bridge BTC L2<>L1');
+command
+    .command('deposit')
+    .description('deposit BTC to l2')
+    .requiredOption('--amount <amount>', 'amount of BTC to deposit', parseFloat)
+    .requiredOption('--receiver-l2-address <receiverL2Address>', 'receiver l2 address')
+    .option('--sender-private-key <senderPrivateKey>', 'sender private key', DEFAULT_DEPOSITOR_PRIVATE_KEY)
+    .option('--network <network>', 'network', DEFAULT_NETWORK)
+    .option('--rpc-url <rcpUrl>', 'RPC URL', DEFAULT_RPC_URL)
+    .option('--rpc-username <rcpUsername>', 'RPC username', DEFAULT_RPC_USERNAME)
+    .option('--rpc-password <rpcPassword>', 'RPC password', DEFAULT_RPC_PASSWORD)
+    .action((cmd: Command) =>
+        deposit(
+            cmd.amount,
+            cmd.receiverL2Address,
+            cmd.senderPrivateKey,
+            cmd.network,
+            cmd.rpcUrl,
+            cmd.rpcUsername,
+            cmd.rpcPassword
+        )
+    );
 
 command
     .command('withdraw')
